+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### New features and improvements
 
 - App: Added new 3:2 and 16:10 forced aspect ratio options.
+- App: Improve full screen frame pacing even further by spin-waiting for up to 1 ms before the frame presentation target.
+- App: Smooth out frame interval adjustments in full screen mode.
 - Debugger: Added basic VDP1 registers inspector window.
 - Settings: Automatically create/suggest a backup RAM file if no path is specified when inserting the cartridge.